+++ library/FilterFormElementChangeEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./FilterFormElementChangeEventDetail.mjs").FilterFormElementChangeEventDetail} FilterFormElementChangeEventDetail */
+
+/**
+ * @typedef {CustomEvent<FilterFormElementChangeEventDetail>} FilterFormElementChangeEvent
+ */