+++ library/FilterFormElementInputEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./FilterFormElementInputEventDetail.mjs").FilterFormElementInputEventDetail} FilterFormElementInputEventDetail */
+
+/**
+ * @typedef {CustomEvent<FilterFormElementInputEventDetail>} FilterFormElementInputEvent
+ */