+++ library/FilterFormElementInputEventDetail.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./InputValue.mjs").InputValue} InputValue */
+
+/**
+ * @typedef {InputValue} FilterFormElementInputEventDetail
+ */