+++ library/FormElementChangeEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./FormElementChangeEventDetail.mjs").FormElementChangeEventDetail} FormElementChangeEventDetail */
+
+/**
+ * @typedef {CustomEvent<FormElementChangeEventDetail>} FormElementChangeEvent
+ */