+++ library/FormElementInputEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./FormElementInputEventDetail.mjs").FormElementInputEventDetail} FormElementInputEventDetail */
+
+/**
+ * @typedef {CustomEvent<FormElementInputEventDetail>} FormElementInputEvent
+ */