+++ library/InputElementChangeEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./InputElementChangeEventDetail.mjs").InputElementChangeEventDetail} InputElementChangeEventDetail */
+
+/**
+ * @typedef {CustomEvent<InputElementChangeEventDetail>} InputElementChangeEvent
+ */