+++ library/InputElementInputEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./InputElementInputEventDetail.mjs").InputElementInputEventDetail} InputElementInputEventDetail */
+
+/**
+ * @typedef {CustomEvent<InputElementInputEventDetail>} InputElementInputEvent
+ */